+++ lib/src/layer/tile_layer/tile_provider/base_tile_provider.dart
@@ -213,9 +213,8 @@ abstract class TileProvider {
   /// When creating a specialized [TileProvider], prefer overriding URL
   /// generation related methods in the following order:
   ///
-  /// 1. [TileLayer.additionalOptions] (or [generateReplacementMap] for more
-  /// advanced usage)
-  /// 2. [populateTemplatePlaceholders]
+  /// 1. [populateTemplatePlaceholders]
+  /// 2. [generateReplacementMap]
   /// 3. [getTileUrl] and/or [getTileFallbackUrl]
   ///
   /// Note to implementors: it is not safe to assume that at least one of
@@ -225,7 +224,7 @@ abstract class TileProvider {
       populateTemplatePlaceholders(
         options.wmsOptions?.getUrl(
               coordinates,
-              options.tileSize,
+              options.tileSize.toInt(),
               options.resolvedRetinaMode == RetinaMode.simulation,
             ) ??
             options.urlTemplate ??
